Refactor and Modularize Text Readers so they can be reused
